feat(sidebar): add arrow key navigation for project/worktree list
