+++ src/renderer/styles/sidebar.css
@@ -52,6 +52,20 @@
   color: var(--text-primary);
 }
 
+.worktree-row:focus {
+  outline: none;
+}
+
+.worktree-row--keyboard-focused {
+  outline: 2px solid var(--accent);
+  outline-offset: -2px;
+}
+
+.project-header--keyboard-focused {
+  outline: 2px solid var(--accent);
+  outline-offset: -2px;
+}
+
 /* Status dot */
 .status-dot {
   width: 8px;
